[004-Encrption-Simple-Interface] 복잡한 Encryption 구조 삭제. 데코레이션 패턴 삭제.
