--- member-context/src/main/java/com/membercontext/memberAPI/application/repository/encryption/javaCrypto/handlers/SHAHandler.java
@@ -1,16 +1,0 @@
-package com.membercontext.memberAPI.application.repository.encryption.javaCrypto.handlers;
-
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.JavaCryptoEncryption;
-import com.membercontext.memberAPI.domain.entity.member.Member;
-
-public class SHAHandler implements JavaCryptoEncryption {
-    @Override
-    public String encrypt(Member member) {
-        return "";
-    }
-
-    @Override
-    public String decrypt(Member member) {
-        return "";
-    }
-}
