--- member-context/src/main/java/com/membercontext/memberAPI/application/repository/encryption/javaCrypto/testFixture/JavaCryptoAdapterTestFixture.java
@@ -1,18 +1,0 @@
-package com.membercontext.memberAPI.application.repository.encryption.javaCrypto.testFixture;
-
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.JavaCryptoAdapter;
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.javaCryptoRepository.JavaCryptoIVRepository;
-
-public class JavaCryptoAdapterTestFixture extends JavaCryptoAdapter {
-
-    public static JavaCryptoAdapter createJavaCryptoAdapter(String key, String algorithm, String mode, String padding, JavaCryptoIVRepository javaCryptoIVRepository) {
-        return JavaCryptoAdapter.builder()
-                .key(key)
-                .algorithm(algorithm)
-                .mode(mode)
-                .padding(padding)
-                .javaCryptoIVRepository(javaCryptoIVRepository)
-                .build();
-    }
-
-}
