--- member-context/src/main/java/com/membercontext/memberAPI/application/repository/encryption/javaCrypto/vo/testFixture/JavaCryptoIVTest.java
@@ -1,7 +1,0 @@
-package com.membercontext.memberAPI.application.repository.encryption.javaCrypto.vo.testFixture;
-
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.vo.JavaCryptoIV;
-
-public class JavaCryptoIVTest extends JavaCryptoIV{
-
-}
