+++ member-context/src/main/java/com/membercontext/memberAPI/application/service/SignUpSerivces/deco/SignUpEmailServiceImpl.java
@@ -1,13 +1,13 @@
-package com.membercontext.memberAPI.application.service.SignUpSerivces;
+package com.membercontext.memberAPI.application.service.SignUpSerivces.deco;
 
+import com.membercontext.memberAPI.application.service.SignUpSerivces.SignUpService;
 import com.membercontext.memberAPI.domain.entity.member.Member;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
-@Service
+//@Service
 @RequiredArgsConstructor
-public class SignUpEmailServiceImpl implements SignUpService{
+public class SignUpEmailServiceImpl implements SignUpService {
 
     private final SignUpService signUpService;
 
@@ -17,6 +17,9 @@ public class SignUpEmailServiceImpl implements SignUpService{
     public String signUp(Member member) {
         signUpService.signUp(member);
         System.out.println("이메일 서비스 추가");
+        //Fixme: 이렇게 데코레이션 패턴으로 추가하는게 테스트 코드를 변경하지 않고 추가하는것 같고
+        //       기능도 추가 제거가 쉽다고 생각해서 이런 구조를 가져갈까 합니다.
+        //       하나 이상 기능 추가는 Configuration을 통해 순서를 정해주려고 합니다.
         return "회원가입 성공";
     }
