+++ member-context/src/main/java/com/membercontext/memberAPI/application/service/SignUpSerivces/deco/SignUpPasswordEncryptionServiceImpl_JavaCrypto.java
@@ -1,14 +1,13 @@
-package com.membercontext.memberAPI.application.service.SignUpSerivces;
+package com.membercontext.memberAPI.application.service.SignUpSerivces.deco;
 
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.JavaCryptoEncryption;
+import com.membercontext.memberAPI.application.service.SignUpSerivces.SignUpService;
 import com.membercontext.memberAPI.domain.entity.member.Member;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
-@Service
+//@Service
 @RequiredArgsConstructor
-public class SignUpPasswordEncryptionServiceImpl_JavaCrypto implements SignUpService{
+public class SignUpPasswordEncryptionServiceImpl_JavaCrypto implements SignUpService {
 
     private final SignUpService signUpService;
