+++ member-context/src/main/java/com/membercontext/memberAPI/domain/entity/member/Member.java
@@ -11,6 +11,7 @@ import lombok.*;
 @Builder(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+//Fixme: 이렇게 Protected로 풀어서 Test Fixture를 만들려고 합니다.
 public class Member {
 
     @Id
