--- member-context/src/main/java/com/membercontext/memberAPI/infrastructure/db/jpa/encription/JavaCryptoIVSpringJPARepository.java
@@ -1,7 +1,0 @@
-package com.membercontext.memberAPI.infrastructure.db.jpa.encription;
-
-import com.membercontext.memberAPI.application.repository.encryption.javaCrypto.vo.JavaCryptoIV;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JavaCryptoIVSpringJPARepository extends JpaRepository<JavaCryptoIV, Long> {
-}
