+++ member-context/src/test/java/com/membercontext/memberAPI/application/service/signUp/deco/integrated/SignUpPasswordEncryptionServiceTest_JavaCrypto.java
@@ -1,13 +1,12 @@
-package com.membercontext.memberAPI.application.service.signUp.integrated;
+package com.membercontext.memberAPI.application.service.signUp.deco.integrated;
 
 import com.membercontext.memberAPI.application.repository.MemberRepository;
-import com.membercontext.memberAPI.application.service.SignUpSerivces.SignUpEmailServiceImpl;
+import com.membercontext.memberAPI.application.service.SignUpSerivces.deco.SignUpPasswordEncryptionServiceImpl_JavaCrypto;
 import com.membercontext.memberAPI.application.service.SignUpSerivces.SignUpService;
 import com.membercontext.memberAPI.application.service.SignUpSerivces.SignUpServiceImpl;
 import com.membercontext.memberAPI.domain.entity.member.Member;
 import com.membercontext.memberAPI.domain.entity.member.testFixture.MemberTest;
 import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,31 +15,35 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Mockito.mock;
+
 
 @SpringBootTest
 @Disabled
-public class SignUpEmailServiceTest {
+class SignUpPasswordEncryptionServiceTest_JavaCrypto {
+
     @Autowired
-    private SignUpService signUpService;
+    private SignUpService sut;
 
     @TestConfiguration
     static class config{
+        //@Autowired
+        //private JavaCryptoAdapter javaCryptoAdapter;
+        //fixme (데코레이션 패턴) 테스트2 : DB와 연결되기에 통합테스트가 필요하다고 생각했고 필요시 하나씩 테스트가 가능합니다.
         @Bean
         @Primary
         public SignUpService signUpService_Test(MemberRepository memberRepository) {
-            return new SignUpEmailServiceImpl(new SignUpServiceImpl(memberRepository));
+            return new SignUpPasswordEncryptionServiceImpl_JavaCrypto(
+                    new SignUpServiceImpl(memberRepository));
         }
     }
 
     @Test
-    @DisplayName("회원가입 성공.")
     void signUp() {
         //given
         Member newMember = MemberTest.createMember(1L, "test", "test", "test", "test", "test", true, 0L);
 
         //when
-        String result = signUpService.signUp(newMember);
+        String result = sut.signUp(newMember);
 
         //then
         assertEquals("회원가입 성공", result);
